feat: hide a file using sys_enter_getdents64 and sys_exit_getdents64, code taken from a tutorial WIP
